Update myScript.js
Modified the state array with a more complex one that holds 3 arrays, each one with 3 objects that contain the keys and the sound bits associated with them. For the moment the sound samples are the ones I recorded for the Simon Game App and are good for testing the app, until I record new ones for the current purpose. Implemented refs for all the sound bits associated with their respective keys, so the mapped keys play the correct sound when clicked. Next I'll need to record a new set of sounds and name them distinctively so the names can be used on displaying the playing sound when clicked.