Update myScript.js
Final update to the project. Recorded two sets of 9 sound effects each, with a distinctive name for each sound that can be referenced to be displayed on the machine's display when played. So each object now has another property "name" that holds each name of the audio bit.
Created a new "Keyboard" React.Component that handles the keyboard events on the specific keys that are found on the app and plays on key press the sound bit associated with that key. Implemented a toggle to chose from 2 library presets of sound effects. The app is fully functional and passes all the test in the freeCodeCamp's tests suite. This is a complete project.